+++ sms/src/main/java/com/tomtom/speedtools/services/SmsModule.java
@@ -18,7 +18,6 @@ package com.tomtom.speedtools.services;
 
 import com.google.inject.Binder;
 import com.google.inject.Module;
-import com.google.inject.Provides;
 import com.google.inject.multibindings.Multibinder;
 import com.tomtom.speedtools.services.sms.SMSDeliveryReportListenerRegistry;
 import com.tomtom.speedtools.services.sms.SMSDeliveryReportResource;
@@ -26,17 +25,8 @@ import com.tomtom.speedtools.services.sms.SMSProviderConnector;
 import com.tomtom.speedtools.services.sms.implementation.SMSDeliveryReportResourceImpl;
 import com.tomtom.speedtools.services.sms.implementation.messagebird.MessageBird;
 import com.tomtom.speedtools.services.sms.implementation.messagebird.MessageBirdProperties;
-import com.tomtom.speedtools.services.sms.implementation.messagebird.MessageBirdResource;
-import com.tomtom.speedtools.services.sms.implementation.messagebird.MessageBirdResourceMock;
 import com.tomtom.speedtools.services.sms.implementation.nexmo.Nexmo;
 import com.tomtom.speedtools.services.sms.implementation.nexmo.NexmoProperties;
-import com.tomtom.speedtools.services.sms.implementation.nexmo.NexmoResource;
-import com.tomtom.speedtools.services.sms.implementation.nexmo.NexmoResourceMock;
-import org.apache.http.impl.client.DefaultHttpClient;
-import org.apache.http.impl.conn.tsccm.ThreadSafeClientConnManager;
-import org.jboss.resteasy.client.ClientExecutor;
-import org.jboss.resteasy.client.ProxyFactory;
-import org.jboss.resteasy.client.core.executors.ApacheHttpClient4Executor;
 
 import javax.annotation.Nonnull;
 import javax.inject.Singleton;
@@ -61,47 +51,4 @@ public class SmsModule implements Module {
         binder.bind(SMSDeliveryReportListenerRegistry.class).in(Singleton.class);
         binder.bind(SMSDeliveryReportResource.class).to(SMSDeliveryReportResourceImpl.class).in(Singleton.class);
     }
-
-    @Provides
-    @Singleton
-    @Nonnull
-    public NexmoResource provideNexmoResource(
-            @Nonnull final NexmoProperties nexmoProperties,
-            @Nonnull final SMSDeliveryReportListenerRegistry smsDeliveryReportListenerRegistry) {
-        assert nexmoProperties != null;
-        assert smsDeliveryReportListenerRegistry != null;
-
-        if (nexmoProperties.isMockEnabled()) {
-            return new NexmoResourceMock(smsDeliveryReportListenerRegistry);
-        }
-
-        // Create an Apache HTTP client executor that has an HTTP client that uses a thread safe connection manager.
-        final ClientExecutor clientExecutor =
-                new ApacheHttpClient4Executor(new DefaultHttpClient(new ThreadSafeClientConnManager()));
-
-        // Return a new Nexmo proxy instance.
-        return ProxyFactory.create(NexmoResource.class, nexmoProperties.getBaseUrl(), clientExecutor);
-    }
-
-    @Provides
-    @Singleton
-    @Nonnull
-    public MessageBirdResource provideMessageBirdResource(
-            @Nonnull final MessageBirdProperties messageBirdProperties,
-            @Nonnull final SMSDeliveryReportListenerRegistry smsDeliveryReportListenerRegistry) {
-        assert messageBirdProperties != null;
-        assert smsDeliveryReportListenerRegistry != null;
-
-        if (messageBirdProperties.isMockEnabled()) {
-            return new MessageBirdResourceMock(smsDeliveryReportListenerRegistry);
-        }
-
-        // Create an Apache HTTP client executor that has an HTTP client that uses a thread safe connection manager.
-        final ClientExecutor clientExecutor =
-                new ApacheHttpClient4Executor(new DefaultHttpClient(new ThreadSafeClientConnManager()));
-
-        // Return a new MessageBird proxy instance.
-        return ProxyFactory.create(MessageBirdResource.class, messageBirdProperties.getBaseUrl(), clientExecutor);
-    }
-
 }