+++ sms/src/main/java/com/tomtom/speedtools/services/sms/implementation/messagebird/MessageBirdResourceMock.java
@@ -19,25 +19,17 @@ package com.tomtom.speedtools.services.sms.implementation.messagebird;
 import com.tomtom.speedtools.services.sms.SMSDeliveryReportListener;
 import com.tomtom.speedtools.services.sms.SMSDeliveryReportListener.DeliveryStatus;
 import com.tomtom.speedtools.services.sms.SMSDeliveryReportListenerRegistry;
-import com.tomtom.speedtools.services.sms.implementation.messagebird.dto.MessageBirdMessageResponse;
-import com.tomtom.speedtools.services.sms.implementation.messagebird.dto.MessageBirdMessageResponse.Item;
-import com.tomtom.speedtools.services.sms.implementation.messagebird.dto.MessageBirdMessageResponse.ResponseCode;
-import org.jboss.resteasy.client.ClientResponse;
-import org.jboss.resteasy.spi.Link;
-import org.jboss.resteasy.spi.LinkHeader;
-import org.jboss.resteasy.util.GenericType;
+import com.tomtom.speedtools.services.sms.implementation.messagebird.dto.MessageBirdResponse;
+import com.tomtom.speedtools.services.sms.implementation.messagebird.dto.MessageBirdResponse.Item;
+import com.tomtom.speedtools.services.sms.implementation.messagebird.dto.MessageBirdResponse.ResponseCode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import javax.ws.rs.core.EntityTag;
+import javax.ws.rs.core.*;
 import javax.ws.rs.core.Link.Builder;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.core.NewCookie;
 import java.lang.annotation.Annotation;
-import java.lang.reflect.Type;
 import java.net.URI;
 import java.util.Date;
 import java.util.Locale;
@@ -61,7 +53,7 @@ public class MessageBirdResourceMock implements MessageBirdResource {
 
 
     @Override
-    public ClientResponse<MessageBirdMessageResponse> sendMessage(
+    public Response sendMessage(
             @Nonnull final String userName,
             @Nonnull final String password,
             @Nullable final Long reference,
@@ -85,9 +77,9 @@ public class MessageBirdResourceMock implements MessageBirdResource {
                         "responseType={}, clientReference={}",
                 status, userName, password, sender, destination, body, responseType, reference);
 
-        final MessageBirdMessageResponse messageBirdMessageResponse = new MessageBirdMessageResponse();
+        final MessageBirdResponse response = new MessageBirdResponse();
         final Item item = new Item();
-        messageBirdMessageResponse.setItem(item);
+        response.setItem(item);
 
         item.setResponseCode(status);
         item.setResponseMessage("Message");
@@ -98,24 +90,12 @@ public class MessageBirdResourceMock implements MessageBirdResource {
         if ((listener != null) && (reference != null)) {
             listener.messageDeliveryReport(reference, DeliveryStatus.DELIVERED);
         }
-
-        return new ClientResponse<MessageBirdMessageResponse>() {
-
-            @Override
-            public MultivaluedMap<String, String> getResponseHeaders() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nonnull
-            public Status getResponseStatus() {
-                return Status.OK;
-            }
+        return new Response() {
 
             @Override
             @Nullable
-            public MessageBirdMessageResponse getEntity() {
-                return messageBirdMessageResponse;
+            public MessageBirdResponse getEntity() {
+                return response;
             }
 
             @Override
@@ -219,76 +199,8 @@ public class MessageBirdResourceMock implements MessageBirdResource {
             }
 
             @Override
-            @Nullable
-            public <T2> T2 getEntity(@Nonnull final Class<T2> type) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nullable
-            public <T2> T2 getEntity(
-                    @Nonnull final Class<T2> type,
-                    @Nonnull final Type genericType) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nullable
-            public <T2> T2 getEntity(
-                    @Nonnull final Class<T2> type,
-                    @Nonnull final Type genericType,
-                    @Nonnull final Annotation[] annotations) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nullable
-            public <T2> T2 getEntity(@Nonnull final GenericType<T2> type) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nullable
-            public <T2> T2 getEntity(@Nonnull final GenericType<T2> type, @Nonnull final Annotation[] annotations) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nonnull
-            public LinkHeader getLinkHeader() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public Link getLocationLink() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            @Nonnull
-            public Link getHeaderAsLink(final String headerName) {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
-            public void resetStream() {
-                // Left empty.
-            }
-
-            @Override
-            public void releaseConnection() {
-                // Left empty.
-            }
-
-            @Override
-            @Nonnull
-            public Map<String, Object> getAttributes() {
-                throw new UnsupportedOperationException();
-            }
-
-            @Override
             public int getStatus() {
-                return getResponseStatus().getStatusCode();
+                return Response.Status.OK.getStatusCode();
             }
 
             @Override